+++ src/main/java/com/lithiumcraft/createresourcegeodes/block/AsurineCatalystBlock.java
@@ -1,5 +1,6 @@
 package com.lithiumcraft.createresourcegeodes.block;
 
+import com.lithiumcraft.createresourcegeodes.block.entity.CatalystBlock;
 import com.lithiumcraft.createresourcegeodes.util.CatalystDataProvider;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;