+++ src/main/java/com/lithiumcraft/createresourcegeodes/block/CrimsiteCatalystBlock.java
@@ -1,11 +1,10 @@
 package com.lithiumcraft.createresourcegeodes.block;
 
+import com.lithiumcraft.createresourcegeodes.block.entity.CatalystBlock;
 import com.lithiumcraft.createresourcegeodes.util.CatalystDataProvider;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
-
-import java.util.Random;
 
 public class CrimsiteCatalystBlock extends CatalystBlock implements CatalystDataProvider {