+++ src/main/java/com/lithiumcraft/createresourcegeodes/sound/ModSounds.java
@@ -15,6 +15,7 @@ public class ModSounds {
 
     public static final Supplier<SoundEvent> CATALYST_BLOCK_TELEPORT = registerSoundEvent("catalyst_block_teleport");
     public static final Supplier<SoundEvent> AGITATOR_INVALID_TIER = registerSoundEvent("agitator_invalid_tier");
+    public static final Supplier<SoundEvent> DONT_BE_A_BOT = registerSoundEvent("dont_be_a_bot");
 
     private static Supplier<SoundEvent> registerSoundEvent(String name) {
         ResourceLocation id = ResourceLocation.fromNamespaceAndPath(CreateResourceGeodes.MOD_ID, name);