+++ src/main/java/com/lithiumcraft/createresourcegeodes/util/CatalystDataProvider.java
@@ -3,6 +3,7 @@ package com.lithiumcraft.createresourcegeodes.util;
 import com.lithiumcraft.createresourcegeodes.config.CatalystShape;
 import com.lithiumcraft.createresourcegeodes.data.CatalystGeneratorDefinition;
 import com.lithiumcraft.createresourcegeodes.registry.ModRegistries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.Block;
@@ -34,28 +35,6 @@ public interface CatalystDataProvider {
         CatalystGeneratorDefinition def = registry.get(getCatalystId());
         return def != null ? def.generatorBlock() : getDefaultGeneratorBlock();
     }
-
-//    default Block getGeneratorBlock(ServerLevel level) {
-//        var registry = level.registryAccess().registryOrThrow(ModRegistries.CATALYST_DEFINITION_KEY);
-//        var id = getCatalystId();
-//
-//        System.out.println("[Catalyst] Looking up: " + id);
-//        if (!registry.containsKey(id)) {
-//            System.out.println("[Catalyst] Registry does NOT contain: " + id);
-//        } else {
-//            System.out.println("[Catalyst] Registry contains: " + id);
-//        }
-//
-//        CatalystGeneratorDefinition def = registry.get(id);
-//
-//        if (def != null) {
-//            System.out.println("[Catalyst] Loaded from datapack: " + id + " -> " + def.generatorBlock());
-//            return def.generatorBlock();
-//        } else {
-//            System.out.println("[Catalyst] Fallback used: " + id);
-//            return getDefaultGeneratorBlock();
-//        }
-//    }
 
     default int getCooldown(ServerLevel level) {
         var registry = level.registryAccess().registryOrThrow(ModRegistries.CATALYST_DEFINITION_KEY);